scripts/requirements.txt
PyPDF2>=3.0.0
pdfplumber>=0.10.0
pymupdf>=1.23.0
python-docx>=1.1.0
openpyxl>=3.1.0
Pillow>=10.0.0
requests>=2.31.0
faiss-cpu>=1.7.4
numpy>=1.24.0
python-dotenv>=1.0.0
